pyproject.toml
[tool.poetry]
name = "src"
version = "0.1.0"
description = ""
authors = [
    "Authors Here"
]

[tool.poetry.dependencies]
python = "^3.7"
djangorestframework = "^3.11.0"
django-cors-headers = "^3.2.1"
django-rest-knox = "^4.1.0"
cloudinary = "^1.20.0"
gunicorn = "^20.0.4"
coverage = "^5.1"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
flake8 = "^3.7.9"

[tool.poetry.scripts]
app = 'src.app:run'
[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"

